requirements.txt
fastapi==0.121.0
uvicorn==0.30.6
ccxt==4.5.15
ta==0.11.0
pandas==2.2.3
python-telegram-bot==21.5
